+++ src/Monitor/MonitorWorkspace.Autorest/Properties/AssemblyInfo.cs
@@ -24,5 +24,5 @@ using System.Runtime.InteropServices;
 [assembly: ComVisible(false)]
 [assembly: CLSCompliant(false)]
 [assembly: Guid("00fd2111-a452-4a86-b971-f439fe047b7a")]
-[assembly: AssemblyVersion("4.6.0")]
-[assembly: AssemblyFileVersion("4.6.0")]
+[assembly: AssemblyVersion("5.0.0")]
+[assembly: AssemblyFileVersion("5.0.0")]