+++ src/CloudService/ChangeLog.md
@@ -18,6 +18,8 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+
+## Version 2.0.0
 * Downgraded the api version of referenced network to 2021-03-01.
 
 ## Version 1.2.0