+++ src/DataFactory/DataFactoryV2/ChangeLog.md
@@ -19,6 +19,8 @@
 -->
 ## Upcoming Release
 * Added support copyComputeScale And pipelineExternalComputeScale in `Set-AzDataFactoryV2IntegrationRuntime` Command
+
+## Version 1.17.1
 * Added ParquetReadSettings in ADF
 * Fixed minor issues