+++ src/Monitor/Monitor/ChangeLog.md
@@ -19,13 +19,29 @@
 -->
 
 ## Upcoming Release
-  * [Breaking Change] Data collection Rule upgraded API version to stable 2022-06-01
-  * [Breaking Change] AMCS removed `Set-AzDataCollectionRule` cmdlet
-  * Added cmdlets for data collection endpoint:
-    * `Get-AzDataCollectionEndpoint`
-    * `New-AzDataCollectionEndpoint`
-    * `Remove-AzDataCollectionEndpoint`
-    * `Update-AzDataCollectionEndpoint`
+
+## Version 5.0.0
+  * [Breaking Change] Action Group upgraded API version to stable 2023-01-01
+  * [Breaking Change] Use new and update cmdlets instead `Set-AzActionGroup` cmdlet
+  * The receiver used subtype cmdlets to create a replacement for command `New-AzActionGroupReceiver`
+    * New-AzActionGroupArmRoleReceiverObject
+    * New-AzActionGroupAutomationRunbookReceiverObject
+    * New-AzActionGroupAzureAppPushReceiverObject
+    * New-AzActionGroupAzureFunctionReceiverObject
+    * New-AzActionGroupEmailReceiverObject
+    * New-AzActionGroupEventHubReceiverObject
+    * New-AzActionGroupItsmReceiverObject
+    * New-AzActionGroupLogicAppReceiverObject
+    * New-AzActionGroupSmsReceiverObject
+    * New-AzActionGroupVoiceReceiverObject
+    * New-AzActionGroupWebhookReceiverObject
+* [Breaking Change] Data collection Rule upgraded API version to stable 2022-06-01
+* [Breaking Change] AMCS removed `Set-AzDataCollectionRule` cmdlet
+* Added cmdlets for data collection endpoint:
+    - `Get-AzDataCollectionEndpoint`
+    - `New-AzDataCollectionEndpoint`
+    - `Remove-AzDataCollectionEndpoint`
+    - `Update-AzDataCollectionEndpoint`
 
 ## Version 4.6.0
 * Fixed `Get-AzInsightsPrivateLinkScope` to support `ResourceId` parameter [#22568]